Move all version-specific code into v1/ and v2/
Leave shared modules at the top level.
